receiving commands ok now
commands sent using the esp32 over curl will be received but will not reply back to the esp32. something to do with the serial port.
curl http://192.168.4.1/api/pose  for example will send a json with id 300 to the arduino, the led will blink, proving the correct command is reaching the arduino, but not coming back to the esp32( per lines 2092-2114 of application function set)

I did comment out the old postracker stuff as well.